✨ feat: GAにサイト内検索イベントを送信
