--- src/types/gtm-event.d.ts
@@ -1,9 +1,0 @@
-type ClickEvent = {
-  action: "click";
-  category: "Other";
-  label: string;
-  value?: string;
-};
-
-export type Event = ClickEvent;
-export default Event;
